Reorganize CLI tool.
